Fix examples and move to .mjs so it can use import syntax
